requirements.txt
fastapi==0.104.1
uvicorn==0.25.0
httpx==0.26.0
sqlalchemy==2.0.23
psycopg2
pydantic
python-dotenv
langchain-community 
langchain-openai
numpy
pydantic-settings
pgvector
pypdf2
langchain-postgres
docling 
aiofiles
redis
python-multipart